refactor: Customer 관련 Dto 패키지 변경
