+++ src/main/java/woowacourse/shoppingcart/dto/customer/CustomerUpdateRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.customer;
 
 import javax.validation.constraints.Max;
 import javax.validation.constraints.NotEmpty;
