+++ src/test/java/woowacourse/shoppingcart/CustomerFixtures.java
@@ -1,8 +1,8 @@
 package woowacourse.shoppingcart;
 
 import woowacourse.shoppingcart.domain.Customer;
-import woowacourse.shoppingcart.dto.CustomerSaveRequest;
-import woowacourse.shoppingcart.dto.CustomerUpdateRequest;
+import woowacourse.shoppingcart.dto.customer.CustomerSaveRequest;
+import woowacourse.shoppingcart.dto.customer.CustomerUpdateRequest;
 
 public class CustomerFixtures {
